docs: restructure basic-usage directory
