+++ docs/cli/basic-usage/example/lib.py
